API-AGENT/requirements.txt
fastapi
uvicorn[standard]
requests
pandas
numpy
tensorflow
python-dotenv
coloredlogs 